Added Assets Class, Added new PRNG, Updated Rg
Assets, such as sprite art and music, will have a static wrapper found in
Assets.java and can therefore be referenced in the form Assets.* from
any source file.

Added a Mersenne Twister implementation to replace builtin Java PRNG.
This class was taken as part of an open-source class from
<http://www.cs.gmu.edu/~sean/research/mersenne/> under the MIT
License. Credits will added to a future Contributers/Thanks file. A static
instance of this PRNG can be found in Rg under the name rng.

Signed-off-by: Alex Colson <guythundar@gmail.com>