Fruit Tiling now working.
This is a huge commit that fixes an outstanding issue of fruits having
incorrect tiling during the boards first generation. This also fixes
some graphical issues and attempts to center the fruits on the screen.
Two-finger tapping will reveal the hidden section of the grid, while
three-finger tapping will generate a new grid.

This commit also changes the icon for the android app in launcher and
the name and verision string. This commit has not been tested on
tablet devices, and may not display correctly. There is also known
issues with targeting x86 android still.


Signed-off-by: Alex Colson <guythundar@gmail.com>