Compiler issues, but code is fine.
There are issues with getting the project to compile correctly, with the
most of the issues relating to a NoClassDefFoundError being thrown.
These errors have been thrown for multiple classes including
org.flixel.FlixelGame and java.awt.Point.

Aside from these compile time issues, the code in this commit reflects
where master and origin/master need to be once the compliation issues
are resolved.

Signed-off-by: Alex Colson <guythundar@gmail.com>