Fixed Compiling Issues, But No Rotation Animation
Everything is being found by the runtime again, so no more missing
classes. With the rework of positioning, animation of Fruit tiles rotating
appears to be broken as well as possibly not swapping in memory. Also
the tiles are now generating incorrectly, having mulitples now being
generated in places that they shouldn't.

Signed-off-by: Alex Colson <guythundar@gmail.com>